refactor(schedule): move schedule lib to repo root
The schedule resolution helpers were under public/js/lib/, where only
the browser could reach them. The notification scheduler is about to
need isScheduledOnDate + effectiveCycles every minute on the server,
and duplicating the logic guarantees drift. Relocate to lib/ as a
single ESM source consumed by both runtimes; the server will pull it
via dynamic import() from CJS in the follow-up.

Browser still fetches /lib/schedule.js: a precise carve-out in the
static handler resolves that one path to the new location, with a
matching exact entry in the public-paths list. No general /lib/*
window, so server-only files (notifications-scheduler.js etc.) stay
behind the auth gate.

Pure refactor; no behaviour change. All 1305 unit tests + 65 e2e
specs pass.

Fixes #396.
