+++ auth/webauthn.js
@@ -134,6 +134,10 @@ function isPublicPath(pathname) {
     '/sw.js',
     '/favicon.ico',
   ];
+  // Exact-path public files. Kept separate from the prefix list so adding
+  // an entry here cannot accidentally open a directory window.
+  const publicExact = ['/lib/schedule.js'];
+  if (publicExact.includes(pathname)) return true;
   return publicPaths.some(p => pathname === p || pathname.startsWith(p));
 }
