+++ lib/schedule.js
@@ -1,8 +1,11 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 // Copyright (C) 2026 Aristocles <https://github.com/Aristocles>
-// public/js/lib/schedule.js
+// lib/schedule.js
 // Pure functions evaluating whether a scheduled item is "on" for a given date.
-// Shared by eh-checklist-card, eh-schedule-timeline, eh-calendar-view etc.
+// Shared between the browser (eh-checklist-card, eh-schedule-timeline,
+// eh-schedule-card, eh-adherence-report, eh-prompt-modal, prompt-queue) and
+// the server (notification scheduler). Stays ESM; CJS callers consume via
+// dynamic import().
 
 const DAY_NAMES_SHORT = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
